requirements.txt
Django==5.2.5
django-allauth==65.9.0
django-unfold==0.60.0
django-storages==1.14.6
python-dotenv==1.0.1
boto3==1.38.36
celery==5.5.3
redis==6.4.0
Pillow==11.2.1
requests==2.32.4
markdown==3.5.1
django-cleanup==8.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.10